Fixed signed/unsigned conversion in las2txt.c
--HG--
extra : convert_revision : svn%3A1766ff46-f334-0410-ab20-d63176f87757/trunk%40865

+++ apps/las2txt.c
@@ -311,7 +311,7 @@ int main(int argc, char *argv[])
                 exit(1);
             }
 
-            len = strlen(file_name_in);
+            len = (int)strlen(file_name_in);
             file_name_out = strdup(file_name_in);
             if (file_name_out[len-3] == '.' && file_name_out[len-2] == 'g' && file_name_out[len-1] == 'z')
             {
